Removed Bukkit and Bungeecord
