+++ src/main/java/net/melion/rgbchat/chat/RGBUtils.kt
@@ -1,6 +1,5 @@
 package net.melion.rgbchat.chat
 
-import net.md_5.bungee.api.ChatColor
 import java.util.regex.Pattern
 
 
@@ -10,13 +9,27 @@ object RGBUtils {
     private val fix3 = Pattern.compile("\\&x[\\&0-9a-fA-F]{12}")
     private val gradient1 = Pattern.compile("<#[0-9a-fA-F]{6}>[^<]*</#[0-9a-fA-F]{6}>")
     private val gradient2 = Pattern.compile("\\{#[0-9a-fA-F]{6}>\\}[^\\{]*\\{#[0-9a-fA-F]{6}<\\}")
-    fun toHexString(red: Int, green: Int, blue: Int): String {
+
+
+    private fun toChatColor(hexCode: String): String {
+        val magic = StringBuilder("§x")
+        val var3: CharArray = hexCode.substring(1).toCharArray()
+        val var4 = var3.size
+
+        for (var5 in 0 until var4) {
+            val c = var3[var5]
+            magic.append('§').append(c)
+        }
+        return magic.toString()
+    }
+
+    private fun toHexString(red: Int, green: Int, blue: Int): String {
         var s = Integer.toHexString((red shl 16) + (green shl 8) + blue)
         while (s.length < 6) s = "0$s"
         return s
     }
 
-    fun applyFormats(textInput: String): String {
+    private fun applyFormats(textInput: String): String {
         var text = textInput
         text = fixFormat1(text)
         text = fixFormat2(text)
@@ -31,7 +44,7 @@ object RGBUtils {
         val m = hex.matcher(text)
         while (m.find()) {
             val hexcode = m.group()
-            text = text.replace(hexcode, ChatColor.of(hexcode).toString())
+            text = text.replace(hexcode, toChatColor(hexcode))
         }
         return text
     }
@@ -42,8 +55,8 @@ object RGBUtils {
     }
 
     //{#RRGGBB}
-    private fun fixFormat2(text: String): String {
-        var text = text
+    private fun fixFormat2(input: String): String {
+        var text = input
         val m = fix2.matcher(text)
         while (m.find()) {
             val hexcode = m.group()
@@ -67,14 +80,14 @@ object RGBUtils {
     }
 
     //<#RRGGBB>Text</#RRGGBB>
-    private fun setGradient1(text: String): String {
-        var text = text
+    private fun setGradient1(input: String): String {
+        var text = input
         val m = gradient1.matcher(text)
         while (m.find()) {
             val format = m.group()
-            val start = IChatBaseComponent.TextColor(format.substring(2, 8))
+            val start = TextColor(format.substring(2, 8))
             val message = format.substring(9, format.length - 10)
-            val end = IChatBaseComponent.TextColor(format.substring(format.length - 7, format.length - 1))
+            val end = TextColor(format.substring(format.length - 7, format.length - 1))
             val applied = asGradient(start, message, end)
             text = text.replace(format, applied)
         }
@@ -82,21 +95,21 @@ object RGBUtils {
     }
 
     //{#RRGGBB>}text{#RRGGBB<}
-    private fun setGradient2(text: String): String {
-        var text = text
+    private fun setGradient2(input: String): String {
+        var text = input
         val m = gradient2.matcher(text)
         while (m.find()) {
             val format = m.group()
-            val start = IChatBaseComponent.TextColor(format.substring(2, 8))
+            val start = TextColor(format.substring(2, 8))
             val message = format.substring(10, format.length - 10)
-            val end = IChatBaseComponent.TextColor(format.substring(format.length - 8, format.length - 2))
+            val end = TextColor(format.substring(format.length - 8, format.length - 2))
             val applied = asGradient(start, message, end)
             text = text.replace(format, applied)
         }
         return text
     }
 
-    private fun asGradient(start: IChatBaseComponent.TextColor, text: String, end: IChatBaseComponent.TextColor): String {
+    private fun asGradient(start: TextColor, text: String, end: TextColor): String {
         val sb = StringBuilder()
         val length = text.length
         for (i in 0 until length) {
